Add sync bar
