+++ src/gui/static/src/app/app.component.ts
@@ -29,27 +29,6 @@ export class AppComponent {
 
   ngOnInit() {
     this.setVersion();
-    IntervalObservable
-      .create(3000)
-      .flatMap(() => this.blockchainService.progress())
-      .takeWhile((response: any) => !response.current || response.current !== response.highest)
-      .subscribe(response => {
-        this.highest = response.highest;
-        this.current = response.current;
-        this.percentage = this.current && this.highest ? (this.current / this.highest * 100) : 0;
-        console.log(response);
-      }, error => console.log(error),
-        () => this.completeLoading());
-  }
-
-  loading() {
-    return !this.current || !this.highest || this.current != this.highest;
-  }
-
-  private completeLoading() {
-    this.current = 999999999999;
-    this.highest = 999999999999;
-    this.walletService.refreshBalances();
   }
 
   private setVersion() {
