+++ src/gui/static/src/app/components/layout/header/header.component.html
@@ -3,13 +3,21 @@
     <app-top-bar [title]="title"></app-top-bar>
     <div class="balance-container">
       <div class="balance">
-        <p class="coins"><span>{{ coins }}</span> sky</p>
-        <p class="dollars">{{ balance }}</p>
+        <p *ngIf="highest && loading" class="coins"><span>{{ percentage | percent:'1.2-2' }}</span></p>
+        <p *ngIf="!loading" class="coins"><span>{{ coins | number:'1.0-6' }}</span> sky</p>
+        <p *ngIf="!loading" class="dollars">{{ balance }}</p>
       </div>
     </div>
     <div class="hour-balance">
-      <p>{{ hours }} SKY Hours</p>
+      <p *ngIf="loading">Syncing blocks {{ current && highest ?  '(' + current + '/'  + highest + ')' : '..' }}</p>
+      <p *ngIf="!loading">{{ hours }} SKY Hours</p>
     </div>
   </div>
   <app-nav-bar></app-nav-bar>
+  <mat-progress-bar
+    *ngIf="loading"
+    class="example-margin"
+    color="primary"
+    mode="determinate"
+    [value]="percentage * 100"></mat-progress-bar>
 </div>
